+++ framework/Furion/DataValidation/Attributes/DataValidationAttribute.cs
@@ -68,11 +68,9 @@ namespace System.ComponentModel.DataAnnotations
                 var resultMessage = dataValidationResult.ValidationResults.FirstOrDefault().ErrorMessage;
 
                 // 进行多语言处理
-                var errorMessage = string.IsNullOrWhiteSpace(ErrorMessage)
-                    ? L.Text == null ? resultMessage : L.Text[resultMessage]
-                    : ErrorMessage;
+                var errorMessage = !string.IsNullOrWhiteSpace(ErrorMessage) ? ErrorMessage : resultMessage;
 
-                return new ValidationResult(string.Format(errorMessage, validationContext.MemberName));
+                return new ValidationResult(string.Format(L.Text == null ? errorMessage : L.Text[errorMessage], validationContext.MemberName));
             }
 
             // 验证成功